wip: CieCardReaderScreen refactoring
