+++ locales/it/index.yml
@@ -648,10 +648,10 @@ authentication:
       content: !include cie/temporary_blocked_cie.md
       button: "Scarica / Apri CieID"
     card:
-      title: Appoggia la carta
+      title: Appoggia la CIE sul retro del dispositivo
       titleiOS: Tieni ferma la carta
       headerTitle: Entra con CIE
-      layCardMessageHeader: Appoggia per qualche secondo la tua carta di identità elettronica al retro del tuo telefono.
+      layCardMessageHeader: Se la lettura non parte, sposta la carta in punti diversi e attendi qualche secondo finché non senti una vibrazione.
       layCardMessageHeaderiOS: Il sensore è posizionato nella parte alta del dispositivo.
       layCardMessageFooter: Se non sai con precisione dove sia l'antenna NFC sposta la carta in punti diversi e attendi qualche secondo finché non senti una vibrazione.
       cieCardValid: Ok, la tua carta CIE è valida! Puoi rimuoverla.
